useless code

+++ baseBuilding.go
@@ -12,7 +12,7 @@ type BaseBuilding struct {
 
 // ConstructionTime returns the duration it takes to build given level
 func (b BaseBuilding) ConstructionTime(level, universeSpeed int64, facilities Facilities, hasTechnocrat, isDiscoverer bool) time.Duration {
-	price := b.GetPrice(int64(level))
+	price := b.GetPrice(level)
 	metalCost := float64(price.Metal)
 	crystalCost := float64(price.Crystal)
 	roboticLvl := float64(facilities.RoboticsFactory)
